requirements.txt
pytest==6.2.5
aws-cdk-lib==2.21.1
constructs>=10.0.0,<11.0.0
opencv-python==4.6.0.66
numpy==1.23.1